+++ src/ft_mlx/utils/button_hooks.c
@@ -6,7 +6,7 @@
 /*   By: stanaka2 <stanaka2@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/06/19 20:32:03 by stanaka2          #+#    #+#             */
-/*   Updated: 2026/06/20 13:39:40 by stanaka2         ###   ########.fr       */
+/*   Updated: 2026/06/23 13:39:13 by stanaka2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -64,13 +64,13 @@ int	button1_motion_hook(int x, int y)
 	delta.y = (y - g_click.y) / PIXEL_PER_ROTATE;
 	if (delta.x != 0)
 	{
-		rotate_camera_yaw((double)delta.x);
+		rotate_camera_yaw((float)delta.x);
 		g_click.x = x;
 		set_draw_flag(true);
 	}
 	if (delta.y != 0)
 	{
-		rotate_camera_pitch((double)-delta.y);
+		rotate_camera_pitch((float)-delta.y);
 		g_click.y = y;
 		set_draw_flag(true);
 	}
